# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Meta-Wall",
    "description": "Meta-Wall API",
    "version": "1.0.0"
  },
  "host": "nodejs-week5-meta-wall.onrender.com",
  "port": "{{PORT}}",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
    "/api/posts/": {
      "get": {
        "tags": [
          "Posts - 搜尋貼文"
        ],
        "description": "搜尋貼文 API",
        "parameters": [
          {
            "name": "timeSort",
            "in": "query",
            "type": "string",
            "description": "ex: desc or monthDesc, default: asc"
          },
          {
            "name": "Keyword",
            "in": "query",
            "type": "string",
            "description": "ex: 測試"
          }
        ],
        "responses": {
          "200": {
            "description": "搜尋成功",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "success": {
                    "type": "boolean",
                    "example": true
                  },
                  "message": {
                    "type": "string",
                    "example": "搜尋成功"
                  },
                  "post": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "662ec324ad9afd608b5c6d88"
                        },
                        "userName": {
                          "type": "string",
                          "example": {
                            "_id": "663272d5917b29ff1a82fac2",
                            "userName": "六角",
                            "photo": "https://thumb.fakeface.rest/thumb_female_21_5a24gdhfjxnhzahsggbfchbsxnjjjjdxzmasnd3jx.jpg"
                          }
                        },
                        "content": {
                          "type": "string",
                          "example": "這是一個測試貼文。"
                        },
                        "image": {
                          "type": "string",
                          "example": "http://nexample.com/image.jpg"
                        },
                        "createdAt": {
                          "type": "string",
                          "example": "2024-02-28T21:44:04.703Z"
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "搜尋失敗",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "success": {
                    "type": "boolean",
                    "example": false
                  },
                  "message": {
                    "type": "string",
                    "example": "搜尋異常"
                  },
                  "errMsg": {
                    "type": "string",
                    "example": "errMsg"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Posts - 新增貼文"
        ],
        "description": "新增貼文 API",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "資料格式",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "userName": {
                  "type": "string",
                  "example": "六角"
                },
                "content": {
                  "type": "string",
                  "example": "這是一篇前端貼文"
                },
                "image": {
                  "type": "string",
                  "example": "http://nexample.com/image.jpg"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "新增成功",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "success": {
                    "type": "boolean",
                    "example": true
                  },
                  "message": {
                    "type": "string",
                    "example": "新增成功"
                  },
                  "post": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "662ec324ad9afd608b5c6d88"
                        },
                        "userName": {
                          "type": "string",
                          "example": "663272d5917b29ff1a82fac2"
                        },
                        "content": {
                          "type": "string",
                          "example": "這是一個測試貼文。"
                        },
                        "image": {
                          "type": "string",
                          "example": "http://nexample.com/image.jpg"
                        },
                        "createdAt": {
                          "type": "string",
                          "example": "2024-02-28T21:44:04.703Z"
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "新增失敗",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "success": {
                    "type": "boolean",
                    "example": false
                  },
                  "message": {
                    "type": "string",
                    "example": "新增貼文失敗"
                  },
                  "errMsg": {
                    "type": "string",
                    "example": "Post validation failed: content: Content 未填寫, userName: 貼文姓名未填寫"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    }
  }
}